fix: Update CMake error message for clarity in Setup.py refactor: Include Model.hpp in Components.hpp feat: Add TransformComponent to entity creation in Entity.cpp

+++ Src/Core/Event/Event.hpp
@@ -176,6 +176,6 @@ namespace Motion::Core
     };
 
 
-    using ApplicationCallbackFunction = std::function<void(uint32_t, IEvent&)>;
+    using ApplicationCallbackFunction = std::function<void(uint64_t, IEvent&)>;
     #define EVENT_CALLBACK(CALLBACK_FUNC) [this](auto&&... args) -> decltype(auto) { return this->CALLBACK_FUNC(std::forward<decltype(args)>(args)...); }
 }
